+++ app/templates/core_components.py
@@ -45,15 +45,18 @@ def sidebar_component(name, subtitle):
     return sidebar
 
 
-def content_params_component(periods):
+def content_params_component(database, all_class_names):
+
+    # Get all the classes
+    sorted_keys = [x['class_name'] for x in database.scan()['Items']]
+    sorted_keys.sort()
+    #https://stackoverflow.com/questions/63502062/python-dash-refresh-page-not-updating-source-data
+
     content_params = html.Div([
         dbc.Row(
             dbc.Col(
                 html.Div(
-                    [
-                        html.H3("Select Class Period", className="display-7"),
-                        # html.P("This will be the class roster that will be grouped..."),
-                    ],
+                    html.H3("Select Class Period", className="display-7"),
                 ),
                 width={"size": 10, "offset": 0},
                 style={"color":"#d3d3d3"}
@@ -61,34 +64,23 @@ def content_params_component(periods):
         ),
         dbc.Row(
             dbc.Col(
-                dbc.RadioItems(
+                dbc.Select(
                     id="period_selection",
-                    className="btn-group",
-                    inputClassName="btn-check",
-                    labelClassName="btn btn-outline-primary",
-                    labelCheckedClassName="active",
-                    options=[
-                        {"label":f"Period {x.split('_')[1]}", "value":x} # +2 b/c periods start at 2
-                    for i, x in enumerate(periods)],
-                    value=periods[0],
+                    options=[{'value':x, 'label':all_class_names[x]} for x in sorted_keys],
                 ),
-                width={'size': 12, 'offset': 0}
+                width={'size': 8, 'offset': 0}
             ),
         ),
         html.Br(),
         dbc.Row(
             dbc.Col(
                 html.Hr(),
-                width={"size": 10, "offset": 0}
             )
         ),
         dbc.Row(
             dbc.Col(
                 html.Div(
-                    [
-                        html.H3("Select Group Size", className="display-7"),
-                        # html.P("This value will be the number of students in each group..."),
-                    ],
+                    html.H3("Select Group Size", className="display-7"),
                 ),
                 width={"size": 10, "offset": 0},
                 style={"color":"#d3d3d3"}
@@ -109,16 +101,12 @@ def content_params_component(periods):
         dbc.Row(
             dbc.Col(
                 html.Hr(),
-                width={"size": 10, "offset": 0}
             )
         ),
         dbc.Row(
             dbc.Col(
                 html.Div(
-                    [
-                        html.H3("Distribute Non-Complete Groups", className="display-7"),
-                        # html.P("This demarcates what to do if the roster isn't cleanly divisible by the group count..."),
-                    ],
+                    html.H3("Distribute Non-Complete Groups", className="display-7"),
                 ),
                 width={"size": 10, "offset": 0},
                 style={"color":"#d3d3d3"}
@@ -148,19 +136,14 @@ def content_roster_component(student_roster_list):
         dbc.Row(
             dbc.Col(
                 html.Div(
-                    [
-                        html.H3("Update Student Attendence", className="display-7"),
-                        # html.P("Students not present today will not be factored into the grouping algorithm..."),
-                    ],
+                    html.H3("Update Student Attendence", className="display-7"),
                 ),
-                width={"size": 10, "offset": 0},
                 style={"color":"#d3d3d3"},
             )
         ),
         dbc.Row(
             dbc.Col(
                 html.Hr(),
-                width={"size": 10, "offset": 0}
             )
         ),
         html.Br(),
@@ -178,11 +161,11 @@ def roster_builder(student_roster_list):
     roster = [dbc.Row([
                 dbc.Col(
                     daq.BooleanSwitch(id=f"{x}", color="#4682b4", on=True),
-                    width={"size": 4, "offset": 0, "align":"start"},
+                    width={"offset": 0, "align":"start"},
                 ),
                 dbc.Col(
                     html.P(f"{x}"),
-                    width={"size": 4, "offset": 0, "justify":"start"},
+                    width={"offset": 0, "justify":"start"},
                 )
             ]) for x in student_roster_list]
     return roster
@@ -194,20 +177,16 @@ def content_table_component():
                 html.Div(
                     [
                         html.H3("Student Groups Table", className="display-7"),
-                        # html.P("Come back to page to select another RANDOM order... "),
                     ],
                 ),
-                width={"size": 10, "offset": 0},
                 style={"color":"#d3d3d3"},
             )
         ),
         dbc.Row(
             dbc.Col(
                 html.Hr(),
-                width={"size": 12, "offset": 0}
             )
         ),
-        # html.Br(),
         html.Div(
             id="grouper_table"
         )],