+++ app/templates/styles.py
@@ -6,21 +6,46 @@ SIDEBAR_STYLE = {
     "bottom": 0,
     "width": "20rem",
     "padding": '8rem 2rem',
-   # "background-color": "#f8f9fa",
 }
 
 # Main content block - on
 CONTENT_STYLE_ON = {
-    "margin-left": "22rem",
+    "position": "fixed",
+    "margin-left": "20rem",
     "margin-right": "2rem",
-    "padding": "2rem 0rem",
     "display": "block",
+    "top": 0,
+    "left": 0,
+    "bottom": 0,
+    "width": "55rem",
+    "padding": '2rem 2rem',
+    "overflow-y": 'scroll'
 }
 
 # Main content block - off
-CONTENT_STYLE_OFF= {
-    "margin-left": "22rem",
+CONTENT_STYLE_OFF = {
+    "position": "fixed",
+    "margin-left": "20rem",
     "margin-right": "2rem",
-    "padding": "2rem 0rem",
     'display': 'none',
+    "top": 0,
+    "left": 0,
+    "bottom": 0,
+    "width": "55rem",
+    "padding": '2rem 2rem',
+    "overflow-y": 'scroll'
+}
+
+# Roster Adjust - on
+OFFCANVAS_ON_STYLE = {
+    'margin': '10px',
+    'width': '95%',
+    'display':'block',
+}
+
+# Roster Adjust - off
+OFFCANVAS_OFF_STYLE = {
+    'margin': '10px',
+    'width': '95%',
+    'display':'None'
 }