+++ CHANGELOG.md
@@ -10,11 +10,18 @@ FEATURES, ENHANCEMENTS, BUG FIXES, DOCS, BREAKING CHANGES, NOTES
 ```
 
 ## Unreleased
+
+### v0.3.0
+### ENHANCEMENTS
+- Added Dynamodb for user to make their own updates to the class roster
+- Added extra dash capablites with offcanvas to do above
+
+### v0.2.0
 ### ENHANCEMENTS
 - Removed NAT Gateway to save money, replaced with endpoint (running)
 - Added scheduled cloudwatch events + lambda to start/stop cluster (running)
 
-##v0.1.0 
+## v0.1.0 
 ### ENHANCEMENTS
 - Updated infa for better ecs management and adding an ALB
 - Updated the make file so it isn't dumb